💄  ui: 다지인 수정 + 폴더/컴포넌트 분리
